requirements.txt
fastapi
uvicorn
openai>=1.0.0
python-dotenv
discord.py
PyNaCl
pytz
Pillow
google-api-python-client
google-auth
google-auth-oauthlib>=0.4.6
google-auth-httplib2
starlette
requests